requirements.txt
fastapi
uvicorn
psycopg2
SQLAlchemy
python-dotenv
alembic
pyodbc
PyJWT